requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.4.2
gitdb==4.0.9
GitPython==3.1.24
huggingface-hub==0.2.1
idna==3.3
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.6
numpy==1.22.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.1
sacremoses==0.0.46
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
streamlit==1.3.1
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
textwrap3==0.9.2
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
